chore: migrate CI from GitLab to GitHub Actions and refactor EditorJS tools
- Replace .gitlab-ci.yml + .gitlab/ with .github/workflows/ci.yml
- GitHub Actions pipeline: format, lint, typecheck, unit tests, SonarCloud,
  Trivy FS/image scans on PRs; release (changeset + ECR push + manifest update) on main
- Refactor EditorJS question blocks: split ShortAnswer/LongAnswer into
  QuestionTitleTool, ShortAnswerTool, LongAnswerTool (hidden from toolbox) and
  ShortAnswerCompositeTool, LongAnswerCompositeTool (visible in toolbox)
- Fix docker-compose ports to match documented values (5174, 3100, 27018)
- Add CLAUDE.md living documentation
- Remove GitLab remote reference from docs
