--- .gitlab-ci.yml
@@ -1,32 +1,0 @@
-default:
-  tags:
-    - local-proxmox
-  cache:
-    key: ${CI_COMMIT_REF_SLUG}
-
-variables:
-  NODE_ENV: test
-  NPM_CONFIG_CACHE: .npm
-  NPM_CONFIG_PRODUCTION: "false"   
-
-before_script:
-  - npm ci
-  - npm ls --depth=0
-
-stages:
-  - code-quality
-  - unit-test
-  - code-analysis
-  - security
-  - release
-
-workflow:
-  rules:
-    - if: '$CI_PIPELINE_SOURCE == "merge_request_event"'
-    - if: '$CI_PIPELINE_SOURCE == "push"'
-
-include:
-  - local: .gitlab/ci/base-jobs.yml
-  - local: .gitlab/ci/sonarqube.yml
-  - local: .gitlab/ci/security-jobs.yml
-  - local: .gitlab/ci/release-jobs.yml
