--- .gitlab/ci/base-jobs.yml
@@ -1,36 +1,0 @@
-format:
-  stage: code-quality
-  resource_group: code-quality
-  script:
-    - npm run format
-  rules:
-    - if: '$CI_PIPELINE_SOURCE == "merge_request_event"'
-
-
-lint:
-  stage: code-quality
-  resource_group: code-quality
-  script:
-    - npm run lint
-  rules:
-    - if: '$CI_PIPELINE_SOURCE == "merge_request_event"'
-
-typecheck:
-  stage: code-quality
-  resource_group: code-quality
-  script:
-    - npm run tsc
-  rules:
-    - if: '$CI_PIPELINE_SOURCE == "merge_request_event"'
-
-unit_test:
-  stage: unit-test
-  image: node:18
-  script:
-    - npm run test:coverage
-  artifacts:
-    paths:
-      - backend/coverage/lcov.info
-      - client/coverage/lcov.info
-  rules:
-    - if: '$CI_PIPELINE_SOURCE == "merge_request_event"'
