--- .gitlab/ci/release-jobs.yml
@@ -1,163 +1,0 @@
-version:
-  stage: release
-  image: node:18-alpine
-  resource_group: release
-  before_script:
-    - apk add --no-cache jq git
-    - npm ci
-  script:
-    # Skip if no pending changesets
-    - echo "VERSION=" > version.env
-
-    - |
-      if [ -z "$(find .changeset -name '*.md' ! -name 'README.md')" ]; then
-        echo "❌ No unreleased changesets. Skipping version bump."
-        echo "VERSION=" > version.env
-        exit 0
-      fi
-
-    # Run version bump
-    - echo "✅ Bumping version using changeset..."
-    - npx @changesets/cli version
-
-    # Get new version
-    - VERSION=$(jq -r '.version' client/package.json)
-    - echo "VERSION=$VERSION" > version.env
-
-    # Git setup
-    - git config --global user.email "ci@example.com"
-    - git config --global user.name "CI Bot"
-
-    - git add .
-    - |
-      if git diff --staged --quiet; then
-        echo "❌ No changes to commit after version bump. Skipping."
-        exit 0
-      fi
-    # Commit version bump and tag
-    - |
-      git commit -m "chore(release): v$VERSION [skip ci]"
-    - |
-      if git tag -l "v$VERSION" | grep -q "v$VERSION"; then
-        echo "⚠️  Tag v$VERSION already exists. Skipping tag creation."
-      else
-        echo "✅ Creating tag v$VERSION"
-        git tag v$VERSION
-      fi
-    - git remote set-url origin
-      https://gitlab-ci-token:${GITLAB_TOKEN}@${CI_SERVER_HOST}/${CI_PROJECT_PATH}.git
-    - git push origin HEAD:main --tags
-  artifacts:
-    reports:
-      dotenv: version.env
-    paths:
-      - ./**/package.json
-      - CHANGELOG.md
-      - version.env
-    expire_in: 1 day
-  rules:
-    - if: '$CI_COMMIT_BRANCH == "main"'
-
-docker_push:
-  stage: release
-  resource_group: release
-  needs:
-    - job: version
-  image: docker:28.1.1
-  services:
-    - name: docker:28.1.1-dind
-      alias: docker
-  dependencies:
-    - version
-  variables:
-    DOCKER_TLS_CERTDIR: ""
-  before_script:
-    - apk add --no-cache curl bash aws-cli jq
-    - echo "Assuming project-specific IAM Role..."
-    - |
-      CREDS=$(aws sts assume-role \
-        --role-arn "arn:aws:iam::283773803331:role/forminit-role" \
-        --role-session-name gitlab-ci-session)
-    - export AWS_ACCESS_KEY_ID=$(echo "$CREDS" | jq -r '.Credentials.AccessKeyId')
-    - export AWS_SECRET_ACCESS_KEY=$(echo "$CREDS" | jq -r '.Credentials.SecretAccessKey')
-    - export AWS_SESSION_TOKEN=$(echo "$CREDS" | jq -r '.Credentials.SessionToken')
-    - echo "Logging into AWS ECR with assumed role credentials..."
-    - aws ecr get-login-password | docker login --username AWS --password-stdin $ECR_REGISTRY
-
-  script:
-    # Guard clause: skip if no version bump
-    - source version.env
-    - |
-      if [ -z "$VERSION" ]; then
-        echo "❌ No version bump detected, skipping Docker push."
-        exit 0
-      fi
-    - echo "Using version $VERSION"
-    - |
-      docker build -f ./client/Dockerfile.prod \
-      --build-arg VITE_GOOGLE_CLIENT_ID="$VITE_GOOGLE_CLIENT_ID" \
-      --build-arg VITE_BACKEND_URL="$VITE_BACKEND_URL" \
-      --build-arg NODE_ENV="production" \
-      -t $ECR_REGISTRY/forminit-client:$VERSION .
-    - docker build -f ./backend/Dockerfile.prod -t $ECR_REGISTRY/forminit-backend:$VERSION .
-    - docker push $ECR_REGISTRY/forminit-client:$VERSION
-    - docker push $ECR_REGISTRY/forminit-backend:$VERSION
-  rules:
-    - if: '$CI_COMMIT_BRANCH == "main"'
-
-update_manifests:
-  stage: release
-  resource_group: release
-  needs:
-    - job: version
-    - job: docker_push
-  image: alpine:latest
-  dependencies:
-    - version
-    - docker_push
-  variables:
-    MANIFEST_REPO_URL: "https://gitlab-ci-token:${GITLAB_TOKEN}@gitlab.com/forminit/forminit-manifests.git"
-    MANIFEST_REPO_BRANCH: "main"
-  before_script:
-    - apk add --no-cache git yq sed
-    - git config --global user.email "ci@example.com"
-    - git config --global user.name "CI Bot"
-  script:
-    # Guard clause: skip if no version bump
-    - source version.env
-    - |
-      if [ -z "$VERSION" ]; then
-        echo "❌ No version bump detected, skipping manifest update.";
-        exit 0;
-      fi
-
-    - echo "✅ Updating manifests for version $VERSION"
-
-    # Clone the manifest repository
-    - git clone $MANIFEST_REPO_URL manifest-repo
-    - cd manifest-repo
-    - git checkout $MANIFEST_REPO_BRANCH
-
-    # Update frontend image tag
-    - |
-      yq eval '.frontend.image.tag = "'$VERSION'"' -i charts/forminit/values.yaml
-
-    # Update backend image tag
-    - |
-      yq eval '.backend.image.tag = "'$VERSION'"' -i charts/forminit/values.yaml
-
-    # Commit and push changes
-    - git add .
-    - |
-      if git diff --staged --quiet; then
-        echo "❌ No changes to commit in Helm chart."
-        exit 0
-      fi
-
-    - |
-      git commit -m "chore(release): update image tag v$VERSION [skip ci]"
-    - git push origin $MANIFEST_REPO_BRANCH
-    - echo "✅ Successfully updated Helm chart with version $VERSION"
-
-  rules:
-    - if: '$CI_COMMIT_BRANCH == "main"'
