--- .gitlab/ci/sonarqube.yml
@@ -1,52 +1,0 @@
-sonarqube-analysis:
-  stage: code-analysis
-  image: sonarsource/sonar-scanner-cli:latest
-  script:
-    - >
-      sonar-scanner
-      -Dsonar.organization=vishnumohanan404
-      -Dsonar.projectKey=vishnumohanan404_forminit
-      -Dsonar.sources=backend/src,client/src,packages/types
-      -Dsonar.tests=backend/src/tests,client/src/tests
-      -Dsonar.exclusions=**/tests/**,**/node_modules/**,**/dist/**,**/build/**,**/*.min.js,**/*.d.ts,**/coverage/**
-      -Dsonar.javascript.lcov.reportPaths=backend/coverage/lcov.info,client/coverage/lcov.info
-      -Dsonar.host.url=https://sonarcloud.io
-      -Dsonar.ws.timeout=300
-      -Dsonar.scanner.socketTimeout=300
-      -Dsonar.scanner.responseTimeout=300
-      -Dhttp.connectionTimeout=60000
-  dependencies:
-    - unit_test
-  rules:
-    - if:
-        '$CI_PIPELINE_SOURCE == "merge_request_event" && ($CI_MERGE_REQUEST_TARGET_BRANCH_NAME ==
-        "dev" || $CI_MERGE_REQUEST_TARGET_BRANCH_NAME == "main")'
-      when: always
-    - when: never
-
-.sonarqube-quality-gate-check:
-  stage: code-analysis
-  image: node:18-alpine
-  script:
-    - apk add --no-cache curl jq
-    - echo "⏳ Waiting 15s for SonarCloud to finish analysis..."
-    - sleep 15
-    - >
-      sh -c '
-        STATUS=$(curl -s -u "$SONAR_TOKEN:" "https://sonarcloud.io/api/qualitygates/project_status?projectKey=vishnumohanan404_forminit" | jq -r ".projectStatus.status");
-        echo "Sonar Quality Gate Status: $STATUS";
-        if [ "$STATUS" != "OK" ]; then
-          echo "❌ Sonar Quality Gate failed. Blocking merge request.";
-          exit 1;
-        fi
-      '
-  needs:
-    - job: sonarqube-analysis
-      artifacts: false
-  rules:
-    - if:
-        '$CI_PIPELINE_SOURCE == "merge_request_event" && ($CI_MERGE_REQUEST_TARGET_BRANCH_NAME ==
-        "dev" || $CI_MERGE_REQUEST_TARGET_BRANCH_NAME == "main")'
-      when: always
-    - when: never
-  allow_failure: false
