+++ docker-compose.yml
@@ -4,7 +4,7 @@ services:
       context: .
       dockerfile: ./client/Dockerfile
     ports:
-      - "5173:5173"
+      - "5174:5173"
     networks:
       - forminit
     env_file:
@@ -19,7 +19,7 @@ services:
       context: .
       dockerfile: ./backend/Dockerfile
     ports:
-      - "3000:3000"
+      - "3100:3000"
     networks:
       - forminit
     depends_on:
@@ -36,7 +36,7 @@ services:
     networks:
       - forminit
     ports:
-      - "27017:27017"
+      - "27018:27017"
     volumes:
       - mongo-data:/data/db
